+++ gizmos/pep_tools/pep_cornerpin_matrix.py
@@ -341,6 +341,19 @@ def gizmo_link(node=None):
 # --------------------------------------------------------------------------- #
 # Help
 # --------------------------------------------------------------------------- #
+_FOOTER_HTML = ('<span style="color:#8a8a8a">Pixel Eye Pictures</span>'
+                '&nbsp;&nbsp;|&nbsp;&nbsp;'
+                '<a href="https://github.com/PixelEyePictures/pep_nuke_tools" '
+                'style="color:#7aa2f7;text-decoration:none">GitHub</a>')
+
+
+def _pep_footer():
+    lbl = QtWidgets.QLabel(_FOOTER_HTML)
+    lbl.setOpenExternalLinks(True)
+    lbl.setAlignment(QtCore.Qt.AlignRight)
+    return lbl
+
+
 def _show_help(parent, title, html):
     dlg = QtWidgets.QDialog(parent)
     dlg.setWindowTitle(title)
@@ -432,6 +445,7 @@ class CornerPinMatrixDialog(QtWidgets.QDialog):
         btns.addWidget(self.cancel_btn)
         btns.addWidget(self.apply_btn)
         lay.addLayout(btns)
+        lay.addWidget(_pep_footer())
 
         self.bake_cb.toggled.connect(self._sync_range)
         self.help_btn.clicked.connect(